examples/postgres_source/pyproject.toml
[project]
name = "postgres-source"
version = "0.1.0"
description = "Demonstrate how to use Postgres tables as the source for CocoIndex."
requires-python = ">=3.11"
dependencies = [
    "cocoindex[embeddings]>=0.2.3",
    "python-dotenv>=1.0.1",
    "pgvector>=0.4.1",
    "psycopg[binary,pool]",
    "numpy",
]

[tool.setuptools]
packages = []
